+++ modules/analyzer/analyzer.py
@@ -3,7 +3,7 @@ import imutils
 import numpy as np
 from .winline import Winline
 from .visualization import Visualization
-from .common import create, find, analyze, define_roi
+from .lib import create, find, analyze, define_roi
 from math import floor, ceil
 from typing import List
 from shapely import affinity
